workflows/qwen3_pdf/pyproject.toml
[project]
name = "qwentest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "pymupdf>=1.24.0",
    "pillow>=12.0.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.57.1",
    "openai>=2.6.1",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
